Add remove to trash

+++ apps/mail/pages/MailIndex.jsx
@@ -2,14 +2,14 @@ import { MailList } from "../cmps/MailList.jsx"
 import { mailsService } from "../services/mail.service.js"
 import { MailFilter } from "../cmps/MailFilter.jsx"
 import { MailCompose } from "../cmps/MailCompose.jsx"
-
+import { MailFolderList } from "../cmps/MailFolderList.jsx"
 const { useState, useEffect } = React
 
 export function MailIndex() {
 
     const [mails, setMails] = useState([])
     const [filterBy, setFilterBy] = useState(mailsService.getDefaultFilter())
-    const [isCompose,setIsCompose]= useState(false)
+    const [isCompose, setIsCompose] = useState(false)
 
     useEffect(() => {
         loadMails()
@@ -29,22 +29,36 @@ export function MailIndex() {
     }
 
     function removeMail(mailId) {
-        mailsService.remove(mailId)
-            .then(() => {
-                setMails(prevMails => prevMails.filter(mail => mailId !== mail.id))
+        mailsService.get(mailId)
+            .then(mail => {
+                if (mail.removedAt === null) {
+                    mail.removedAt = Date.now()
+                    mailsService.save(mail)
+                        .then(savedMail =>
+                            setMails(prevMails =>
+                                prevMails.filter(currmail => currmail.id !== savedMail.id))
+                        )
+                } else {
+                    mailsService.remove(mailId)
+                        .then(() => {
+                            setMails(prevMails => prevMails.filter(mail => mailId !== mail.id))
+                        })
+                }
             })
+
     }
 
-    function onSetIsCompose(){
+    function onSetIsCompose() {
         setIsCompose(!isCompose)
     }
 
 
     return <section className="mails-container">
-        <MailFilter filterBy={filterBy} onSetFilterBy={onSetFilterBy}/>
+        <MailFilter filterBy={filterBy} onSetFilterBy={onSetFilterBy} />
         <button className='compose-btn' onClick={onSetIsCompose}><i className="fa-solid fa-pencil"></i>Compose</button>
-        <MailList mails={mails} onRemove={removeMail}/>
-       {isCompose && <MailCompose onSetIsCompose={onSetIsCompose}/>}
+        <MailList mails={mails} onRemove={removeMail} />
+        <MailFolderList filterBy={filterBy} onSetFilterBy={onSetFilterBy} />
+        {isCompose && <MailCompose onSetIsCompose={onSetIsCompose} />}
     </section>
 }
